requirements.txt
# Core FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database and ORM
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
greenlet==3.0.3

# Data validation and settings
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0.post1
validators==0.22.0

# Authentication and security
python-jose[cryptography]==3.3.1
passlib[bcrypt]==1.7.4
cryptography==41.0.8
bcrypt==4.1.2

# File handling
python-multipart==0.0.6
aiofiles==23.2.1
pillow==10.1.0
python-magic==0.4.27
PyPDF2==3.0.1
python-docx==1.1.0

# Redis and background tasks
redis==5.0.1
celery==5.3.4
kombu==5.3.4
vine==5.1.0

# AWS and cloud services
boto3==1.34.34

# AI and ML
openai==1.6.1
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2

# Environment and configuration
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3

# HTTP client and WebSocket
httpx==0.26.0
websockets==12.0

# Utilities
python-slugify==8.0.1
Jinja2==3.1.2

# Monitoring
psutil==5.9.6
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
coverage==7.3.4

# Development tools
black==23.12.1
flake8==6.1.0
mypy==1.8.0
# Add to requirements-dev.txt
pre-commit==3.6.0
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.26.0  # For testing FastAPI
factory-boy==3.3.0  # For test data generation
# Add to requirements-prod.txt  
sentry-sdk[fastapi]==1.39.2  # Error tracking
prometheus-client==0.19.0     # Metrics
